fix: rename Todo -> ITodo
