+++ Frontend/src/interfaces/todo.ts
@@ -1,4 +1,4 @@
-export interface Todo {
+export interface ITodo {
   id: number;
   text: string;
   completed: boolean;
@@ -8,5 +8,5 @@ export interface TodoService {
   addTodo: (text: string) => void;
   toggleTodo: (id: number) => void;
   deleteTodo: (id: number) => void;
-  getTodos: () => Todo[];
+  getTodos: () => ITodo[];
 }
