+++ Frontend/src/services/TodoService.ts
@@ -1,10 +1,10 @@
-import { Todo, TodoService } from "../types/todo";
+import { ITodo, TodoService } from "../interfaces/todo";
 
 class TodoServiceImpl implements TodoService {
-  private todos: Todo[] = [];
+  private todos: ITodo[] = [];
 
   addTodo = (text: string): void => {
-    const newTodo: Todo = {
+    const newTodo: ITodo = {
       id: Date.now(),
       text,
       completed: false,
@@ -23,9 +23,9 @@ class TodoServiceImpl implements TodoService {
     this.todos = this.todos.filter((todo) => todo.id !== id);
   };
 
-  getTodos = (): Todo[] => {
+  getTodos = (): ITodo[] => {
     return this.todos;
   };
 }
 
-export const todoService: TodoService = new TodoServiceImpl()
+export const todoService: TodoService = new TodoServiceImpl();
